% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotatedIntronRanges.gencode.v19}
\alias{annotatedIntronRanges.gencode.v19}
\title{The intron ranges annotated with the promoter information.}
\format{A GRanges object of 344,651 ranges with metadata. The intron ranges
  annotated with the promoter information.}
\usage{
annotatedIntronRanges.gencode.v19
}
\description{
A GRanges object. The intron ranges annotated with the promoter information.
}
\keyword{datasets}
