% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction-read-count.R
\name{calculatePromoterReadCounts}
\alias{calculatePromoterReadCounts}
\title{Calculate the promoter read counts using junction read counts approach for
all the input junction files}
\usage{
calculatePromoterReadCounts(exonRanges, intronRanges,
  junctionFilePaths = NULL, junctionFileLabels = NULL,
  junctionType = "tophat", numberOfCores = 1)
}
\arguments{
\item{exonRanges}{A GRanges object containing reduced exon ranges by gene}

\item{intronRanges}{A Granges object containing the annotated unique intron
ranges. These ranges willbe used for counting the reads}

\item{junctionFilePaths}{A character vector. The list of junction files for
which the junction read counts will be calculated}

\item{junctionFileLabels}{A character vector. The labels of junction files
for which the junction read counts will be calculated. These labels will be
used as column names for the output data.frame object}

\item{junctionType}{A character. Type of the junction bed file, either
'tophat'(default) or 'star'}

\item{numberOfCores}{A numeric value. The number of cores to be used for
counting junction reads. Defaults to 1 (no parallelization). This parameter
will be used argument to mclapply function hence require 'parallel' package
to be installed}
}
\value{
A data.frame object. The number of junction reads per promoter (rows)
  for each sample (cols)
}
\description{
Calculate the promoter read counts using junction read counts approach for
all the input junction files
}
\examples{
\dontrun{
junctionFilePaths <- c('./sample1-tophat.bed', './sample2-tophat.bed')
junctionFileLabels <- c('sample1', 'sample2')
promoterReadCounts <- calculatePromoterReadCounts(exonReducedRanges,
                                                   intronRanges.annotated,
                                                   junctionFilePaths,
                                                   junctionFileLabels,
                                                   junctionType = 'tophat',
                                                   numberOfCores = 1)
}

}
