% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-promoter-activity.R
\name{getGeneExpression}
\alias{getGeneExpression}
\title{Prepare the gene expression table including the gene ids}
\usage{
getGeneExpression(absolutePromoterActivity)
}
\arguments{
\item{absolutePromoterActivity}{data.frame of absolute promoter activity with promoter and gene ids}
}
\value{
data.frame of gene expression with gene ids
}
\description{
Prepare the gene expression table including the gene ids
}
\examples{
\dontrun{
geneExpression <- getGeneExpression(absolutePromoterActivity)
}

}
