% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-promoter-activity.R
\name{getRelativePromoterActivity}
\alias{getRelativePromoterActivity}
\title{Prepare the relative promoter activity table including the promoter and gene
ids}
\usage{
getRelativePromoterActivity(absolutePromoterActivity, geneExpression)
}
\arguments{
\item{absolutePromoterActivity}{data.frame of absolute promoter activity with
promoter and gene ids}

\item{geneExpression}{data.frame of gene expression with gene ids}
}
\value{
data.frame of relative promoter activity with promoter and gene ids
}
\description{
Prepare the relative promoter activity table including the promoter and gene
ids
}
\examples{

## absolutePromoterActivity is an object returned from getAbsolutePromoterActivity
## geneExpression is an object returned from getGeneExpression
absolutePromoterActivity <- readRDS(system.file('extdata/testdata/tophat2', 
                                                'absolutePromoterActivity.rds', 
                                                 package = 'proActiv'))
geneExpression <- readRDS(system.file('extdata/testdata/tophat2', 
                                        'geneExpression.rds', 
                                         package = 'proActiv'))
relativePromoterActivity <- getRelativePromoterActivity(absolutePromoterActivity,
                                                        geneExpression)

}
