% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{getUnannotatedReducedExonRanges}
\alias{getUnannotatedReducedExonRanges}
\title{Get the reduced first exon ranges without metadata for the annotation txdb
provided}
\usage{
getUnannotatedReducedExonRanges(txdb, species = "Homo_sapiens",
  numberOfCores = 1)
}
\arguments{
\item{txdb}{A txdb object. The txdb object of the annotation version for
which promoters will be identified}

\item{species}{A character object. The genus and species of the organism to
be used in keepStandardChromosomes(). Supported species can be seen with
names(genomeStyles()).}

\item{numberOfCores}{A numeric value. The number of cores to be used for
reducing first exons of each gene. Defaults to 1 (no parallelization). This
parameter will be used argument to mclapply function hence require
'parallel' package to be installed}
}
\value{
A GRanges object. The reduced first exon ranges for each promoter
}
\description{
Get the reduced first exon ranges without metadata for the annotation txdb
provided
}
\examples{
\dontrun{
exonReducedRanges <- getUnannotatedReducedExonRanges(txdb,
                                                     species = 'Homo_sapiens',
                                                     numberOfCores = 1)
}

}
