% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction-read-count.R
\name{normalizePromoterReadCounts}
\alias{normalizePromoterReadCounts}
\title{Normalize promoter read counts using DESeq2}
\usage{
normalizePromoterReadCounts(promoterReadCounts)
}
\arguments{
\item{promoterReadCounts}{A data.frame object. The number of junction reads
per promoter (rows) for each sample (cols)}
}
\value{
A data.frame object. The normalized number of junction reads per
  promoter (rows) for each sample (cols) using DESeq2 counts function.
  Requires 'DESeq2' package to be installed
}
\description{
Normalize promoter read counts using DESeq2
}
\examples{

## promoterReadCounts is an object returned from calculatePromoterReadCounts
promoterReadCounts <- readRDS(system.file('extdata/testdata/tophat2',
                                          'promoterCounts.rds', 
                                           package = 'proActiv'))
normalizedPromoterReadCounts <- normalizePromoterReadCounts(
                                           promoterReadCounts)

}
