% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-proActiv.R
\name{plotPromoters}
\alias{plotPromoters}
\title{Wrapper function returning Summarized Experiment object giving promoter 
counts and activity}
\usage{
plotPromoters(
  result,
  gene,
  txdb,
  ranges,
  cex.title = 0.9,
  cex.axis = 0.9,
  cex.main = 1,
  blk.width = 500,
  blk.fill = "grey",
  blk.border = "darkgrey",
  label.col = "black",
  label.size = 0.7,
  arrow.width = NULL,
  arrow.fill = "transparent",
  arrow.border = "grey"
)
}
\arguments{
\item{result}{A SummarizedExperiment object with assays giving promoter 
counts and activity with gene expression stored as column data and 
promoter gene id mapping stored as row data}

\item{gene}{A character vector of length 1. Single gene of interest to 
be plotted}

\item{txdb}{A TxDb object. The txdb must correspond to the genome version 
used in running proActiv. Here, it is recommended to use the same txdb in 
generating promoter annotations}

\item{ranges}{A list of GRanges. Each entry in the list should correspond to 
a transcript that will be visualized, with Genomic Ranges giving the exons 
corresponding to that transcript}

\item{cex.title}{A numeric value. Size of axis labels. Defaults to 0.9}

\item{cex.axis}{A numeric value. Size of axis and axis ticks. Defaults to 0.9}

\item{cex.main}{A numeric value. Size of plot name. Defaults to 1}

\item{blk.width}{A numeric value. The width of promoters blocks in the 
data track. Defaults to 500 (bases)}

\item{blk.fill}{A character vector of length 1. The fill colour of the 
promoter blocks in the data track. Defaults to 'grey'}

\item{blk.border}{A character vector of length 1. The border colour of the 
promoter blocks in the data track. Defaults to 'darkgrey'}

\item{label.col}{A character vector of length 1. The font colour of the 
promoter ID label in the annotation track. Defaults to 'black'}

\item{label.size}{A numeric value. The size of the promoter ID label in the 
annotation track. Defaults to 0.7}

\item{arrow.width}{A numeric value. The width of promoter arrows in the 
annotation track. This value is internally calculated based on the gene 
of interest}

\item{arrow.fill}{A character vector of length 1. The fill colour of the 
promoter arrows in the annotation track. Defaults to 'transparent'}

\item{arrow.border}{A character vector of length 1. The border colour of 
the promoter arrows in the annotation track. Defaults to 'grey'}
}
\value{
Outputs a plot of the promoters of the gene of interest across 
  conditions, along with a model of transcripts belonging to the gene
}
\description{
Wrapper function returning Summarized Experiment object giving promoter 
counts and activity
}
\examples{
 
 gene <- 'ENSG00000076864.19'
 ## Genomic Ranges giving exons by transcripts of gene 
 ranges <- readRDS(system.file('extdata/vignette/annotations', 
                               'exonsBy.rap1gap.rds',
                               package = 'proActiv'))
 ## summarizedExperiment returned by proActiv (subsetted to gene RAP1GAP)
 result <- readRDS(system.file('extdata/vignette/annotations',
                               'result.rap1gap.rds',
                               package ='proActiv'))
 plotPromoters(result = result, gene = gene, ranges = ranges)
  
                           
}
