% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{prepareAnnotatedIntronRanges}
\alias{prepareAnnotatedIntronRanges}
\title{Prepare the annotated intron ranges for the annotating txdb specified}
\usage{
prepareAnnotatedIntronRanges(txdb, species = "Homo_sapiens",
  promoterIdMapping)
}
\arguments{
\item{txdb}{A txdb object. The txdb object of the annotation version for
which promoters will be identified}

\item{species}{A character object. The genus and species of the organism to
be used in keepStandardChromosomes(). Supported species can be seen with
names(genomeStyles()).}

\item{promoterIdMapping}{A data.frame object. The id mapping between
transcript ids, names, TSS ids, promoter ids and gene ids}
}
\value{
A GRanges object. The intron ranges annotated with the promoter
  information.
}
\description{
Prepare the annotated intron ranges for the annotating txdb specified
}
\examples{
\dontrun{
intronRanges.annotated <- prepareAnnotatedIntronRanges(txdb,
                                                       species = 'Homo_sapiens',
                                                       promoterIdMapping)
}

}
