% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{prepareAnnotatedReducedExonRanges}
\alias{prepareAnnotatedReducedExonRanges}
\title{Prepare the reduced exon ranges for each promoter using the annotation txdb
specified}
\usage{
prepareAnnotatedReducedExonRanges(txdb, species = "Homo_sapiens",
  promoterIdMapping, exonReducedRanges)
}
\arguments{
\item{txdb}{A txdb object. The txdb object of the annotation version for
which promoters will be identified}

\item{species}{A character object. The genus and species of the organism to
be used in keepStandardChromosomes(). Supported species can be seen with
names(genomeStyles()).}

\item{promoterIdMapping}{A data.frame object. The id mapping between
transcript ids, names, TSS ids, promoter ids and gene ids}

\item{exonReducedRanges}{A GRanges object. The reduced first exon ranges for
each promoter without metadata}
}
\value{
A GRanges object. The reduced first exon ranges for each promoter
  with promoter metadata
}
\description{
Prepare the reduced exon ranges for each promoter using the annotation txdb
specified
}
\examples{
\dontrun{
exonReducedRanges <- prepareAnnotatedReducedExonRanges(txdb,
                                                       species = 'Homo_sapiens',
                                                       promoterIdMapping,
                                                       exonReducedRanges)
}

}
