% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{preparePromoterAnnotationData}
\alias{preparePromoterAnnotationData}
\title{Prepare promoter annotation data for the user specified txdb object}
\usage{
preparePromoterAnnotationData(txdb, species = "Homo_sapiens",
  numberOfCores = 1)
}
\arguments{
\item{txdb}{A txdb object. The txdb object of the annotation version for
which promoters will be identified.}

\item{species}{A character object. The genus and species of the organism to
be used in keepStandardChromosomes(). Supported species can be seen with
names(genomeStyles()).}

\item{numberOfCores}{A numeric value. The number of cores to be used for
reducing first exons of each gene. Defaults to 1 (no parallelization). This
parameter will be used argument to mclapply function hence require
'parallel' package to be installed.}
}
\value{
A PromoterAnnotation object. The reduced exon ranges, annotated
  intron ranges, promoter coordinates and the promoter id mapping are
  attributes of the promoter annotation data.
}
\description{
Prepare promoter annotation data for the user specified txdb object
}
\examples{
\dontrun{
promoterAnnotation <- preparePromoterAnnotationData(txdb,
                                                    species = 'Homo_sapiens',
                                                    numberOfCores = 1)
}
}
