% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{preparePromoterCoordinates}
\alias{preparePromoterCoordinates}
\title{Retrieve promoter coordinates (TSS) for each promoter with gene id and internal promoter state}
\usage{
preparePromoterCoordinates(exonReducedRanges, promoterIdMapping)
}
\arguments{
\item{exonReducedRanges}{A GRanges object. The reduced first exon ranges for
each promoter without metadata}

\item{promoterIdMapping}{A data.frame object. The id mapping between
transcript ids, names, TSS ids, promoter ids and gene ids}
}
\value{
A GRanges object. Promoter coordinates (TSS) with gene id and internal promoter state
}
\description{
Retrieve promoter coordinates (TSS) for each promoter with gene id and internal promoter state
}
\examples{
\dontrun{
promoterCoordinates <- preparePromoterCoordinates(exonReducedRanges,
                                                          promoterIdMapping)
}
}
