% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{preparePromoterIdMapping}
\alias{preparePromoterIdMapping}
\title{Prepare the id mapping between transcript ids, names, TSS ids, promoter ids
and gene ids}
\usage{
preparePromoterIdMapping(txdb, species = "Homo_sapiens",
  exonReducedRanges)
}
\arguments{
\item{txdb}{A txdb object. The txdb object of the annotation version for
which promoters will be identified}

\item{species}{A character object. The genus and species of the organism to
be used in keepStandardChromosomes(). Supported species can be seen with
names(genomeStyles()).}

\item{exonReducedRanges}{A GRanges object. The reduced first exon ranges for
each promoter}
}
\value{
A data.frame object. The id mapping between transcript ids, names,
  TSS ids, promoter ids and gene ids
}
\description{
Prepare the id mapping between transcript ids, names, TSS ids, promoter ids
and gene ids
}
\examples{
\dontrun{
promoterIdMapping <- preparePromoterIdMapping(txdb,
                                               species = 'Homo_sapiens',
                                               exonReducedRanges)
}

}
