% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoterAnnotation.gencode.v34}
\alias{promoterAnnotation.gencode.v34}
\title{Promoter annotation data for Gencode.v34 including all the annotation objects
required for promoter activity estimation}
\format{
A PromoterAnnotation (S4 Class) object containing all the promoter
  annotation objects for Gencode.v34. The object has 3 slots: \describe{
  \item{intronRanges}{A GRanges object of 383,654 ranges corresponding
  to introns, annotated with the associated transcript.} 
  \item{promoterIdMapping}{The id mapping between transcript names, 
  promoter ids and gene ids for Gencode v34.} 
  \item{promoterCoordinates}{A GRanges object of 122,635 ranges
  showing the tss coordinate for each promoter of Gencode v34,
  annotated with the associated gene id, coordinate of the 3' end of the first
  reduced exon, and intron id.} }
}
\usage{
promoterAnnotation.gencode.v34
}
\description{
A GRanges object containing the tss coordinate for each promoter for Gencode
v34
}
\keyword{datasets}
