% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoterAnnotation.gencode.vM1}
\alias{promoterAnnotation.gencode.vM1}
\title{Promoter annotation data for Gencode.vM1 including all the annotation objects
required for promoter activity estimation}
\format{
A PromoterAnnotation (S4 Class) object containing all the promoter
  annotation objects for Gencode.vM1. The object has 3 slots: \describe{
  \item{intronRanges}{A GRanges object of 243,332 ranges corresponding
  to introns, annotated with the associated transcript.} 
  \item{promoterIdMapping}{The id mapping between transcript names, 
  promoter ids and gene ids for Gencode vM1.} 
  \item{promoterCoordinates}{A GRanges object of 60,768 ranges
  showing the tss coordinate for each promoter of Gencode vM1,
  annotated with the associated gene id, coordinate of the 3' end of the first
  reduced exon, and intron id.} }
}
\usage{
promoterAnnotation.gencode.vM1
}
\description{
A GRanges object containing the tss coordinate for each promoter for Gencode
vM1
}
\keyword{datasets}
