% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoterCoordinates.gencode.v19}
\alias{promoterCoordinates.gencode.v19}
\title{Transcription start site coordinates for each promoter identified using
Gencode v19}
\format{A GRanges object of 113,076 ranges showing the tss coordinate for
  each promoter of Gencode v19}
\usage{
promoterCoordinates.gencode.v19
}
\description{
A GRanges object containing the tss coordinate for each promoter for Gencode
v19
}
\keyword{datasets}
