% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoterIdMapping.gencode.v19}
\alias{promoterIdMapping.gencode.v19}
\title{The id mapping between transcript ids, names, TSS ids, promoter ids and gene
ids}
\format{A data frame with 196520 rows and 5 variables: \describe{
  \item{transcriptId}{the transcript ids of all transcripts for Gencode v19}
  \item{transcriptName}{the transcript names of all transcripts for Gencode
  v19} \item{tssId}{the custom tss ids of each transcript for Gencode v19}
  \item{promoterId}{the custom promoter ids of each transcript for Gencode
  v19} \item{geneId}{the gene ids of all transcripts for Gencode v19} }}
\usage{
promoterIdMapping.gencode.v19
}
\description{
A data.frame object. The id mapping between transcript ids, names, TSS ids,
promoter ids and gene ids
}
\keyword{datasets}
