% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reducedExonRanges.gencode.v19}
\alias{reducedExonRanges.gencode.v19}
\title{Reduced exon ranges for the first exons of each gene to determine the
transcript to promoter mapping for Gencode v19}
\format{A GRanges object of 113,076 reduced exon ranges with promoter
  metadata for Gencode v19}
\usage{
reducedExonRanges.gencode.v19
}
\description{
A GRanges object containing the reduced first exons ranges by gene with
promoter metadata
}
\keyword{datasets}
