% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assays}
\alias{rowData}
\alias{colData}
\alias{metadata}
\alias{as_tibble}
\alias{\%>\%}
\alias{group_by}
\alias{mutate}
\alias{filter}
\alias{n}
\alias{arrange}
\alias{rownames_to_column}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr:reexports]{as_tibble}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:context]{n}}}

  \item{S4Vectors}{\code{\link[S4Vectors:Annotated-class]{metadata}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}}

  \item{tibble}{\code{\link[tibble:rownames]{rownames_to_column}}}
}}

