% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\name{preparePromoterAnnotation}
\alias{preparePromoterAnnotation}
\title{Prepare promoter annotation data for the user specified txdb object}
\usage{
preparePromoterAnnotation(file, species)
}
\arguments{
\item{file}{A character object. The file path to a gtf/gff or txdb of the 
annotation version for which promoters will be identified.}

\item{species}{A character object. The genus and species of the organism to
be used in keepStandardChromosomes(). Supported species can be seen with
names(genomeStyles()).}
}
\value{
A PromoterAnnotation object. The annotated
  intron ranges, promoter coordinates and the promoter id mapping are
  attributes of the promoter annotation data.
}
\description{
Prepare promoter annotation data for the user specified txdb object
}
\examples{
\dontrun{
promoterAnnotation <- preparePromoterAnnotation(file,
                                                species = 'Homo_sapiens')
}
}
