% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proActiv.R
\name{proActiv}
\alias{proActiv}
\title{Wrapper function returning SummarizedExperiment object giving promoter counts and activity}
\usage{
proActiv(
  files,
  promoterAnnotation,
  fileLabels = NULL,
  genome = NULL,
  ncores = 1
)
}
\arguments{
\item{files}{A character vector. The list of input files for 
which the junction read counts will be calculated}

\item{promoterAnnotation}{A PromoterAnnotation object containing the
intron ranges, promoter coordinates and promoter id mapping}

\item{fileLabels}{A character vector. The labels of input files 
for which the junction read counts will be calculated. These labels will be 
used as column names for each output data.frame object. If not provided,
filenames will be used as labels. Defaults to NULL}

\item{genome}{A character. Genome version. Must be specified if input file
type is a BAM file. Defaults to NULL}

\item{ncores}{A numeric value. The number of cores to be used for 
counting junction reads. Defaults to 1 (no parallelization). This parameter 
will be used as an argument to BiocParallel::bplapply}
}
\value{
A SummarizedExperiment object with assays giving promoter counts and activity 
  with gene expression stored as column data and promoter gene id mapping stored as 
  row data
}
\description{
Wrapper function returning SummarizedExperiment object giving promoter counts and activity
}
\examples{
\dontrun{
files <- c('./sample1-tophat.bed', './sample2-tophat.bed')
proActivFromJunctions <- proActiv(files = files,
                                  promoterAnnotation = promoterAnnotation, 
                                  fileLabels = NULL, 
                                  genome = NULL, 
                                  ncores = 1)

files <- c('./sample1.bam', './sample2.bam')
proActivFromBAM <- proActiv(files = files,
                            promoterAnnotation  = promoterAnnotation,
                            fileLabels = NULL,
                            genome = 'hg19',
                            ncores = 1)}
                           
}
