% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PromoterAnnotation-class.R
\docType{class}
\name{PromoterAnnotation-class}
\alias{PromoterAnnotation-class}
\alias{PromoterAnnotation}
\alias{reducedExonRanges,PromoterAnnotation-method}
\alias{annotatedIntronRanges,PromoterAnnotation-method}
\alias{promoterIdMapping,PromoterAnnotation-method}
\alias{promoterCoordinates,PromoterAnnotation-method}
\alias{reducedExonRanges<-,PromoterAnnotation-method}
\alias{annotatedIntronRanges<-,PromoterAnnotation-method}
\alias{promoterIdMapping<-,PromoterAnnotation-method}
\alias{promoterCoordinates<-,PromoterAnnotation-method}
\title{S4 class for promoter annotation data for a specific annotation version}
\usage{
PromoterAnnotation(reducedExonRanges = GRanges(),
  annotatedIntronRanges = GRanges(), promoterIdMapping = data.frame(),
  promoterCoordinates = GRanges())

reducedExonRanges(x)

\S4method{reducedExonRanges}{PromoterAnnotation}(x)

annotatedIntronRanges(x)

\S4method{annotatedIntronRanges}{PromoterAnnotation}(x)

promoterIdMapping(x)

\S4method{promoterIdMapping}{PromoterAnnotation}(x)

promoterCoordinates(x)

\S4method{promoterCoordinates}{PromoterAnnotation}(x)

reducedExonRanges(x) <- value

\S4method{reducedExonRanges}{PromoterAnnotation}(x) <- value

annotatedIntronRanges(x) <- value

\S4method{annotatedIntronRanges}{PromoterAnnotation}(x) <- value

promoterIdMapping(x) <- value

\S4method{promoterIdMapping}{PromoterAnnotation}(x) <- value

promoterCoordinates(x) <- value

\S4method{promoterCoordinates}{PromoterAnnotation}(x) <- value
}
\arguments{
\item{reducedExonRanges}{A GRanges object containing reduced exon ranges}

\item{annotatedIntronRanges}{A GRanges object containing annotated intron ranges}

\item{promoterIdMapping}{A data.frame containing mapping between transcript, tss, promoter and gene ids}

\item{promoterCoordinates}{A GRanges object containing promoter coordinates}

\item{x}{A PromoterAnnotation object}

\item{value}{User specified value to set}
}
\description{
S4 class for promoter annotation data for a specific annotation version

Constructor for PromoterAnnotation class

Getter for the reduced exon ranges

Getter for the annotated intron ranges

Getter for the promoter id mapping

Getter for the promoter coordinates

Setter for the reduced exon ranges

Setter for the annotated intron ranges

Setter for the promoter id mapping

Setter for the promoter coordinates
}
\section{Slots}{

\describe{
\item{\code{reducedExonRanges}}{A GRanges object. The reduced first exon ranges for
each promoter without metadata.}

\item{\code{annotatedIntronRanges}}{A GRanges object. The intron ranges annotated with the promoter
information.}

\item{\code{promoterIdMapping}}{A data.frame object. The id mapping between transcript ids, names,
TSS ids, promoter ids and gene ids.}

\item{\code{promoterCoordinates}}{A GRanges object. Promoter coordinates (TSS) with gene id and internal promoter state}
}}

